Add BackPressedCallbackHelper to unify back pressed logic
